+++ src/Controllers/PageController.php
@@ -8,43 +8,16 @@ use KirchenImWeb\Helpers\ParameterChecker;
 use Psr\Container\ContainerInterface;
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Http\Message\ResponseInterface as Response;
-use Railto\TwigAssetVersionExtension;
-use Slim\Views\Twig;
-use Slim\Views\TwigExtension;
-use Twig_Extensions_Extension_I18n;
 
 /**
  * Class PageController
  *
  * @package KirchenImWeb\Controllers
  */
-class PageController {
-    private $twig;
+class PageController extends TwigController {
 
     public function __construct(ContainerInterface $container) {
-        // Configure Twig for templates and translations.
-        $this->twig = new Twig(__DIR__ . '/../templates', [
-            'cache' => (defined('DEBUG') && DEBUG) ? false : __DIR__ . '/../../cache'
-        ]);
-
-        // Instantiate and add Slim specific extension
-        $basePath = rtrim(str_ireplace('index.php', '', $container['request']->getUri()->getBasePath()), '/');
-        $this->twig->addExtension(new TwigExtension($container['router'], $basePath));
-
-        // Add asset extension for Twig.
-	    $this->twig->addExtension(new TwigAssetVersionExtension(__DIR__. '/../../assets/rev-manifest.json'));
-
-        // Add I18n extension for Twig and init textdomain and set default language.
-	    $this->twig->addExtension(new Twig_Extensions_Extension_I18n());
-        $textdomain = "kirchen-im-web";
-        bindtextdomain($textdomain, 'lang');
-        bind_textdomain_codeset($textdomain, 'UTF-8');
-        textdomain($textdomain);
-
-	    // Pass global variables to the view.
-	    $this->twig->offsetSet('domain', $_SERVER['HTTP_HOST']);
-	    $this->twig->offsetSet('currentPath', $container['request']->getUri()->getPath());
-	    $this->twig->offsetSet('config', Configuration::getInstance());
+        parent::__construct($container);
     }
 
     public function index(Request $request, Response $response, array $args) {
@@ -174,12 +147,6 @@ class PageController {
             'headline' => _('Seite nicht gefunden'),
             'text' => _('Leider konnte die gewünschte Seite nicht gefunden werden. Möglicherweise wurde dieser Eintrag gelöscht.')
         ])->withStatus(404);
-    }
-
-    public function sitemap(Request $request, Response $response, array $args) {
-        return $this->twig->render($response, 'sitemap.xml.twig', [
-            'ids' => Database::getInstance()->getIds()
-        ])->withHeader('Content-Type', 'text/xml; charset=UTF-8');
     }
 
 	public function opensearch(Request $request, Response $response, array $args) {